Did some reorganization of functions for better readability. Deleted some unused imports
